requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
cryptography==41.0.4
Werkzeug==2.3.7
humanize==4.8.0
python-dotenv==1.0.0